refactor: déplacement du fetch dans un service
